chore: rename Core to core
