+++ src/ComponentMap.ts
@@ -1,4 +1,4 @@
-import { PaymentMethodsResponse } from './Core/types';
+import { PaymentMethodsResponse } from './core/types';
 
 /**
  * Find payment method in JSON response or \paymentMethods API
