+++ src/core/configuration.ts
